requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-experimental==0.0.57
langchain-google-genai==1.0.2
langchain-text-splitters==0.0.1
langsmith==0.1.47
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.17.1
orjson==3.10.0
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pyngrok==7.1.6
pyparsing==3.1.2
python-decouple==3.8
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tenacity==8.2.3
tqdm==4.66.2
twilio==9.0.4
typing-inspect==0.9.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
